server/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Flask==3.0.3
Flask-Cors==5.0.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.1
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==5.4.1
requests==2.32.3
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
wadllib==1.3.6
Werkzeug==3.0.4
zipp==1.0.0
